+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Arinoto changelog
 
+## 2022-11-11 [feature]  Cluster together things of same class in diagrams
+
+The idea is to have APIs (and maybe more) grouped visually together in
+the diagram. Also nice to have them on the left side of the diagram.
+[More information](<spec/cluster-vis/design.md>)
+
 ## 2022-10-31 [feature]  Ticketing within repo
 
 Design specifications and ticketing integrated in the project’s
@@ -7,4 +13,4 @@ repository.
 [More information](<spec/000 ticketing/design.md>)
 
 ### Modules changed:
- - @arinoto/vis 1.2.0->1.2.1
+ - @arinoto/vis 1.2.0->1.3.0